requirements.txt
flask
binance-connector
gunicorn
gspread
oauth2client